fix: file I/O race conditions causing data corruption
- Always flush memory-mapped view before unmapping on Windows (SetEndOfFile error)

- Add msync before munmap on Linux/macOS (JSON file truncation)

- Fix CreateFileMapping error check (NULL, not INVALID_HANDLE_VALUE)

- Use GetFileSizeEx/SetFilePointerEx for 64-bit file size support

- Fix GetSize() return type from unsigned int to size_t

- Clamp library Load() to buffer size to prevent overflow

- Add proper cleanup on all file open failure paths

+++ fileio.h
@@ -37,7 +37,7 @@ class File {
     return fp_;
   }
 
-  unsigned int GetSize() const {
+  size_t GetSize() const {
     return size_;
   }
